+++ src/app/_components/chatting/chat.module.scss
@@ -3,12 +3,7 @@
   flex-direction: column;
   align-items: flex-end;
 }
-.chatScroll {
-  overflow-y: auto;
-  max-height: 60vh;
-}
-.chatUserNameFalse {
-}
+
 .chatHistory {
   width: 150px;
   background-color: rgb(255, 216, 222);
@@ -25,7 +20,4 @@
 .date {
   margin-bottom: 15px;
   font-size: 12px;
-}
-.ChatInputSpace {
-  display: flex;
 }
