+++ src/app/_components/chatting/chatInput.module.scss
@@ -1,0 +1,3 @@
+.ChatInputSpace {
+  display: flex;
+}
