+++ src/app/_components/chatting/chatList.module.scss
@@ -1,0 +1,4 @@
+.chatScroll {
+  overflow-y: auto;
+  max-height: 60vh;
+}
